pyproject.toml
[project]
name = "mkhuda-blog-llm-rag"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12,<5.0",
    "faiss-cpu>=1.12,<1.13",
    "langchain-community>=0.3,<0.4",
    "langchain-core>=0.3,<0.4",
    "langchain-openai>=0.2,<0.3",
    "mysql-connector-python>=9.0,<10.0",
    "openai>=1.52,<2.0",
    "pandas>=2.2,<3.0",
    "python-dotenv>=1.0,<2.0",
    "tqdm>=4.66,<5.0",
]
